Fix rubocop linters
